Added distinct ID and user ID back
